requirements.txt
certifi==2018.4.16
chardet==3.0.4
Django==1.9.9
django-embed-video==1.1.2
django-js-asset==1.1.0
django-mptt==0.9.0
django-tinymce==2.7.0
idna==2.6
mysqlclient==1.3.12
Pillow==5.1.0
requests==2.18.4
urllib3==1.22
